Remove some PdfViewer warnings.
Fix build.

git-svn-id: http://skia.googlecode.com/svn/trunk@12438 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ experimental/PdfViewer/pdfparser/native/SkPdfNativeDoc.cpp
@@ -379,7 +379,7 @@ SkPdfNativeObject* SkPdfNativeDoc::readObject(int id/*, int expectedGeneration*/
     const unsigned char* current = fFileContent + startOffset;
     const unsigned char* end = fFileContent + fContentLength;
 
-    SkPdfNativeTokenizer tokenizer(current, end - current, fAllocator, this);
+    SkPdfNativeTokenizer tokenizer(current, (int) (end - current), fAllocator, this);
 
     SkPdfNativeObject idObj;
     SkPdfNativeObject generationObj;
@@ -439,7 +439,7 @@ void SkPdfNativeDoc::fillPages(SkPdfPageTreeNodeDictionary* tree) {
         return;
     }
 
-    int cnt = kids->size();
+    int cnt = (int) kids->size();
     for (int i = 0; i < cnt; i++) {
         SkPdfNativeObject* obj = resolveReference(kids->objAtAIndex(i));
         if (fMapper->mapPageObjectDictionary(obj) != kPageObjectDictionary_SkPdfNativeObjectType) {
